Add meta class usage and algo basic methods
Implemented the usage of the meta class to navigate pages and completed
the methods used to extract Stops and RouteStopPatterns in a BBox. Fixed
a bug in which if the string used by the Json deserializer wasn't coded
in UTF-8 it would throw an Exception when parsing some unicode
characters.